redo param parsing and allow selection of # of AZs
this required redoing the divideSubnet function to allow
dynamically changing the depth calculated. It must be the same
as the largest level used when selecting parts of AZs.
To split into N AZs, level log2(numAZs) is used. Then, it is
necessary to select three levels further:

* private
* public
* protected

Therefore, the depth needed is log2(numAZs) + 3.